Ajout Point/Line et remplacement int en double

+++ Vector/Vector2D.cpp
@@ -7,7 +7,7 @@ Vector2D::Vector2D(Vector2D const& v) {
 	this->_x = v._x;
 	this->_y = v._y;
 }
-Vector2D::Vector2D(int x, int y) {
+Vector2D::Vector2D(double x, double y) {
 	_instance++;
 	this->_x = x;
 	this->_y = y;
@@ -18,16 +18,16 @@ Vector2D::~Vector2D() {
 unsigned int Vector2D::getNumberOfInstance() {
 	return _instance;
 }
-int Vector2D::getX() const {
+double Vector2D::getX() const {
 	return this->_x;
 }
-int Vector2D::getY() const {
+double Vector2D::getY() const {
 	return this->_y;
 }
-void Vector2D::setX(int x) {
+void Vector2D::setX(double x) {
 	this->_x = x;
 }
-void Vector2D::setY(int y) {
+void Vector2D::setY(double y) {
 	this->_y = y;
 }
 std::string Vector2D::toString() const {
@@ -37,7 +37,7 @@ bool Vector2D::isEqual(Vector2D const & v) const {
 	return (this->_x == v._x && this->_y == v._y);
 }
 double Vector2D::normalize() const {
-	return std::sqrt((double)_x*_x + _y*_y);
+	return std::sqrt(_x*_x + _y*_y);
 }
 
 //:: STATIC ATTRIBUT
