+++ Vector/Vector3D.h
@@ -5,15 +5,15 @@ class Vector3D : public Vector2D
 public:
 	//:: CONSTRUCTEURS
 	Vector3D(Vector3D const& v);
-	Vector3D(int x = 0, int y = 0, int z = 0);
+	Vector3D(double x = 0, double y = 0, double z = 0);
 
 	//:: DESTRUCTEURS
 	~Vector3D();
 
 	//:: GETTEURS - ACCESSEURS
-	int getZ() const;
+	double getZ() const;
 
-	void setZ(int z);
+	void setZ(double z);
 
 	//:: METHODES
 	std::string toString() const;
@@ -22,7 +22,7 @@ public:
 	double normalize() const;
 
 private:
-	int _z;
+	double _z;
 
 };
